+++ processor/src/main/java/react4j/processor/Constants.java
@@ -5,9 +5,9 @@ import java.util.List;
 
 final class Constants
 {
+  static final String DAGGER_BINDS_CLASSNAME = "dagger.Binds";
   static final String DAGGER_SUBCOMPONENT_CLASSNAME = "dagger.Subcomponent";
   static final String DAGGER_MODULE_CLASSNAME = "dagger.Module";
-  static final String DAGGER_PROVIDES_CLASSNAME = "dagger.Provides";
   static final String COMPONENT_CLASSNAME = "react4j.Component";
   static final String PROP_ANNOTATION_CLASSNAME = "react4j.annotations.Prop";
   static final String PROP_DEFAULT_ANNOTATION_CLASSNAME = "react4j.annotations.PropDefault";