+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Update the javascript included within package to react version `16.6.0`.
 * **\[arez\]** Update the `org.realityforge.arez` dependencies to version `0.112`.
+* **\[processor\]** Change the way the Dagger2 factory is defined by switching from a static method using a
+  `@Provide` annotation to using an abstract method using a `@Binds` annotation. This results in more optimized
+  code being output by the dagger compiler.
 
 ### [v0.103](https://github.com/react4j/react4j/tree/v0.103) (2018-10-19)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.102...v0.103)